+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Implement more methods:
+  * `get_apple_pay_auth`
+  * `get_bulk_operations`, `get_bulk_detailed_status`, `get_bulk_operation_status_count`
 * Allow to override the `page_size` in `get_paginated` and all methods that use it
 
 ## 1.10.1 (2024/02/07)